MoleculeMO-master/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64

# Obtained via the terminal command: conda list -e > requirements.txt

_pytorch_select=0.1=cpu_0
beautifulsoup4=4.9.3=pyha847dfd_0
blas=1.0=mkl
brotlipy=0.7.0=py36h2bbff1b_1003
ca-certificates=2021.5.25=haa95532_1
cairo=1.14.12=hf171d8a_3
certifi=2021.5.30=py36haa95532_0
cffi=1.14.5=py36hcd4344a_0
cryptography=3.4.7=py36h71e12ea_0
cudatoolkit=10.2.89=h74a9793_1
cudnn=7.6.5=cuda10.2_0
cycler=0.10.0=py36haa95532_0
dataclasses=0.8=pyh4f3eec9_6
freetype=2.10.4=hd328e21_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.2=pyhd3eb1b0_0
intel-openmp=2021.2.0=haa95532_616
ipynb-py-convert=0.4.6=pyh9f0ad1d_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.1=py36hd77b12b_0
libboost=1.67.0=hd9e427e_4
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_0
matplotlib=3.3.4=py36haa95532_0
matplotlib-base=3.3.4=py36h49ac443_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
ninja=1.10.2=h6d14046_1
numpy=1.19.2=py36hadc3359_0
numpy-base=1.19.2=py36ha3acd2a_0
olefile=0.46=py36_0
openssl=1.1.1k=h2bbff1b_0
pandas=1.1.5=py36hd77b12b_0
pillow=8.2.0=py36h4fa10fc_0
pip=21.1.1=py36haa95532_0
pixman=0.40.0=he774522_0
py-boost=1.67.0=py36h8300f20_4
pycparser=2.20=py_2
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h6538335_2
pysocks=1.7.1=py36haa95532_0
python=3.6.13=h3758d61_0
python-dateutil=2.8.1=pyhd3eb1b0_0
pytorch=1.8.1=py3.6_cuda10.2_cudnn7_0
pytz=2021.1=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
rdkit=2020.03.1.0=py36h3d1ada6_1
scikit-learn=0.24.2=py36hf11a4ad_1
scipy=1.5.2=py36h9439919_0
selenium=3.141.0=py36h2bbff1b_1000
setuptools=52.0.0=py36haa95532_0
sip=4.19.8=py36h6538335_0
six=1.15.0=py36haa95532_0
soupsieve=2.2.1=pyhd3eb1b0_0
sqlite=3.35.4=h2bbff1b_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
torchaudio=0.8.1=py36
torchvision=0.9.1=py36_cu102
tornado=6.1=py36h2bbff1b_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py36haa95532_0
wincertstore=0.2=py36h7fe50ca_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
